py/probe_info_service/app_engine/requirements.txt
flask==1.1.1
google-auth==1.27.0
google-cloud-datastore==1.12.0
google-cloud-logging==1.15.0
gunicorn==20.0.4
protobuf==3.12.2
PyYAML==5.3.1
urllib3==1.24.2
